SiteUser 엔티티 추가, 다른 엔티티의 주키를 Long 타입으로

+++ src/main/java/com/ll/exam/sbb/question/QuestionController.java
@@ -38,7 +38,7 @@ public class QuestionController {
     }
 
     @GetMapping("/detail/{id}")
-    public String detail(Model model, @PathVariable int id) {
+    public String detail(Model model, @PathVariable long id) {
         Question question = questionService.getQuestion(id);
 
         model.addAttribute("question", question);
